requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.0.1
streamlit==1.28.1
python-docx==1.1.0
googletrans==4.0.0rc1
gtts==2.4.0
numpy==1.24.3
matplotlib==3.7.2 